Moved PremisLogger and PremisEventBuilder from metadata to fcrepo-clients module.  Generating event ids in PremisLogger
